Get the right num of taskset-ed cores.
Before, get_physcial would just return the number of cores in the
system independent of whatever we give it; this meant it would spawn
enough threads to compete with the benchmark program that was runnning on the same machine.

We also spawned n threads instead of n-1 threads; this would mean that
the main-thread would have to compete with the threads in the pool.