Minimized Crit. section
In the handle request function; the thread with the mutex lock would
work on the request; and since the request wouldn't timeout untill the
server completes the work, it will block all other threads from doing
work/enter the handle function. In order to counter this, the Crit.
section was moved to where we handle inger requests.